refactored code, now use require
