+++ public/css/adlister-new.css
@@ -11,6 +11,37 @@ button:hover {
 	opacity: 0.7;
 }
 
+.nbar {
+	padding: 10px;
+}
+
+.title {
+	display: inline-block;
+	color: #ddd;
+	font-size: 1.2em;
+}
+
+.search {
+	display: inline-block;
+	margin-left: 15px;
+	margin-right: 15px;
+}
+
+.showUsername {
+	display: inline-block;
+	color: #ddd;
+}
+
+.showPost {
+	background-color: #eee;
+	margin-left: 15px;
+	margin-right: 15px;
+	border: 2px solid #888;
+	border-radius: 5px;
+	width: 100px;
+	outline: none;
+}
+
 .front {
 	position: relative;
 	width: 100%;
@@ -91,6 +122,7 @@ button:hover {
 }
 
 .showLogIn {
+	background-color: #eee;
 	margin-left: 15px;
 	border: 2px solid #888;
 	border-radius: 5px;
