+++ views/ps_main.php
@@ -1,0 +1,22 @@
+<div class="main">
+    <div class="container front">
+        <div class="banner">
+            <img src="/img/placeholder.png" class="bannerImg">
+        </div>
+    </div>
+
+    <div class="container-fluid body">
+        <table>
+            <tr>
+                <td><img src="/img/placeholder-small.png" class="itemsImg"></td>
+                <td><img src="/img/placeholder-small.png" class="itemsImg"></td>
+                <td><img src="/img/placeholder-small.png" class="itemsImg"></td>
+            </tr>
+            <tr>
+                <td><img src="/img/placeholder-small.png" class="itemsImg"></td>
+                <td><img src="/img/placeholder-small.png" class="itemsImg"></td>
+                <td><img src="/img/placeholder-small.png" class="itemsImg"></td>
+            </tr>    
+        </table>
+    </div>
+</div>
