+++ views/ps_navbar.php
@@ -1,0 +1,10 @@
+<div class="nbar">
+    <div class="title">photoSHOP</div>
+    <button class="showLogIn">Log In</button>
+    <button class="showPost">Post</button>
+    <form class="search">
+        <input type="text">
+        <button type="submit">Search</button>
+    </form>
+    <div class="showUsername">Insert Username</div>
+</div>
